# openapi.yaml
openapi: 3.1.0
info:
  title: Take home tech task
  version: v1.0.0
  description: |-
    We want you to create a REST API for Eagle Bank which conforms to this OpenAPI specification which allows a user to create, fetch, update and delete a bank account and deposit or withdraw money from the account. These will be stored as transactions against a bank account which be retrieved but not modified or deleted.
...
paths:
  /v1/accounts:
    post:
      tags:
        - account
      description: Create a new bank account
      operationId: createAccount
      security:
        - bearerAuth: []
      requestBody:
        description: Create a new bank account for the user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBankAccountRequest'
        required: true
      responses:
        '201':
          description: Bank Account has been created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountResponse'
        '400':
          description: Invalid details supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
        '401':
          description: Access token is missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: The user is not allowed to access the transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags:
        - account
      description: List accounts
      operationId: listAccounts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The list of bank accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListBankAccountsResponse'
        '401':
          description: Access token is missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    CreateBankAccountRequest:
      type: object
      required:
        - name
        - accountType
      properties:
        name:
          type: string
          examples:
            - "Personal Bank Account"
            - "My Account"
        accountType:
          type: string
          enum: 
            - "personal"
...
    ListBankAccountsResponse:
      type: object
      required:
        - accounts
      properties:
        accounts:
          type: array
          items:
             $ref: "#/components/schemas/BankAccountResponse"
    BankAccountResponse:
      type: object
      required:
        - accountNumber
        - sortCode
        - name
        - accountType
        - balance
        - currency
        - createdTimestamp
        - updatedTimestamp
      properties:
        accountNumber:
          type: string
          format: ^01\d{6}$
          examples:
            - "01234567"
            - "01765432"
        sortCode:
          type: string
          enum: 
            - "10-10-10"
        name:
          type: string
        accountType:
          type: string
          enum: 
            - "personal"
        balance:
          type: number
          format: double
          minimum: 0.00
          maximum: 10000.00
          description: "Currency amount with up to two decimal places"
          examples:
            - 0.00
            - 1000.00
        currency:
          type: string
          enum:
            - "GBP"
        createdTimestamp:
          type: string
          format: 'date-time'
        updatedTimestamp:
          type: string
          format: 'date-time'
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    BadRequestErrorResponse:
      type: object
      required:
        - message
        - details
      properties:
        message:
          type: string
        details:
          type: array
          items:
            type: object
            required:
              - field
              - message
              - type
            properties:
              field:
                type: string
              message:
                type: string
              type:
                type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT